Oct 24, 2024, 3:36 PM

+++ matriz/teste1.c
@@ -2,7 +2,7 @@
 #include <math.h>
 
 int main(){
-    float matriz[5][2];
+    int matriz[5][2];
     
     for (int i = 0; i < 5; i++){
         for( int j = 0; j < 2; j ++ ){
